+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.14.1 - TBD
 - Fixed issue where `DBNStore.to_df()` would raise an exception if no records were present
+- Fixed exception message when creating a DBNStore from an empty data source
 
 ## 0.14.0 - 2023-06-14
 - Added support for reusing a `Live` client to reconnect